+++ app/src/main/java/eu/kanade/tachiyomi/source/SourceExtensions.kt
@@ -87,7 +87,7 @@ fun Source.isLocalOrStub(): Boolean = isLocal() || this is StubSource
 
 fun AnimeSource.getNameForAnimeInfo(): String {
     val preferences = Injekt.get<SourcePreferences>()
-    val enabledLanguages = preferences.enabledLanguages().get()
+    val enabledLanguages = preferences.enabledAnimeLanguages().get()
         .filterNot { it in listOf("all", "other") }
     val hasOneActiveLanguages = enabledLanguages.size == 1
     val isInEnabledLanguages = lang in enabledLanguages