+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/animesource/AnimeSourcesFilterScreenModel.kt
@@ -3,7 +3,7 @@ package eu.kanade.tachiyomi.ui.browse.animesource
 import androidx.compose.runtime.Immutable
 import cafe.adriel.voyager.core.model.StateScreenModel
 import cafe.adriel.voyager.core.model.screenModelScope
-import eu.kanade.domain.source.interactor.ToggleLanguage
+import eu.kanade.domain.source.interactor.ToggleAnimeLanguage
 import eu.kanade.domain.source.service.SourcePreferences
 import kotlinx.collections.immutable.ImmutableList
 import tachiyomi.core.common.preference.getAndSet
@@ -24,13 +24,13 @@ import uy.kohesive.injekt.api.get
 class AnimeSourcesFilterScreenModel(
     private val animeSourceManager: AnimeSourceManager = Injekt.get(),
     private val preferences: SourcePreferences = Injekt.get(),
-    private val toggleLanguage: ToggleLanguage = Injekt.get(),
+    private val toggleLanguage: ToggleAnimeLanguage = Injekt.get(),
 ) : StateScreenModel<AnimeSourcesFilterScreenModel.State>(State.Loading) {
 
     init {
         screenModelScope.launchIO {
             combine(
-                preferences.enabledLanguages().changes(),
+                preferences.enabledAnimeLanguages().changes(),
                 preferences.disabledAnimeSources().changes(),
             ) { enabledLanguages, disabledSources ->
                 val allSources = animeSourceManager.getCatalogueSources()